+++ scripts/update_poetry.sh
@@ -1,7 +1,6 @@
 #!/bin/sh
-#!/bin/sh
 
 SCRIPT=$(realpath "$0")
 SCRIPTPATH=$(dirname "$SCRIPT")
 APP_ROOT="$(dirname "$SCRIPTPATH")"
-cat $APP_ROOT/requirements.txt | xargs poetry add $APP_ROOT
+cat $APP_ROOT/requirements.txt | xargs poetry add -C $APP_ROOT/envs/